% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{associate_certificate}
\alias{associate_certificate}
\title{Associates an AWS Certificate Manager (ACM) Amazon Resource Name (ARN) with AWS Elemental MediaConvert}
\usage{
associate_certificate(Arn)
}
\arguments{
\item{Arn}{[required] The ARN of the ACM certificate that you want to associate with your MediaConvert resource.}
}
\description{
Associates an AWS Certificate Manager (ACM) Amazon Resource Name (ARN) with AWS Elemental MediaConvert.
}
\section{Accepted Parameters}{
\preformatted{associate_certificate(
  Arn = "string"
)
}
}

