% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{create_job}
\alias{create_job}
\title{Create a new transcoding job}
\usage{
create_job(Role, Settings, BillingTagsSource = NULL,
  ClientRequestToken = NULL, JobTemplate = NULL, Queue = NULL,
  UserMetadata = NULL)
}
\arguments{
\item{Role}{Required. The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html.}

\item{Settings}{}

\item{BillingTagsSource}{}

\item{ClientRequestToken}{Idempotency token for CreateJob operation.}

\item{JobTemplate}{When you create a job, you can either specify a job template or specify the transcoding settings individually}

\item{Queue}{Optional. When you create a job, you can specify a queue to send it to. If you don't specify, the job will go to the default queue. For more about queues, see the User Guide topic at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html.}

\item{UserMetadata}{User-defined metadata that you want to associate with an MediaConvert job. You specify metadata in key/value pairs.}
}
\description{
Create a new transcoding job. For information about jobs and job settings, see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html
}
\examples{

}
