% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{create_job_template}
\alias{create_job_template}
\title{Create a new job template}
\usage{
create_job_template(Settings, Name, Category = NULL, Description = NULL,
  Queue = NULL, Tags = NULL)
}
\arguments{
\item{Settings}{}

\item{Name}{The name of the job template you are creating.}

\item{Category}{Optional. A category for the job template you are creating}

\item{Description}{Optional. A description of the job template you are creating.}

\item{Queue}{Optional. The queue that jobs created from this template are assigned to. If you don't specify this, jobs will go to the default queue.}

\item{Tags}{The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.}
}
\description{
Create a new job template. For information about job templates see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html
}
\examples{

}
