% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{create_preset}
\alias{create_preset}
\title{Create a new preset}
\usage{
create_preset(Settings, Name, Category = NULL, Description = NULL,
  Tags = NULL)
}
\arguments{
\item{Settings}{}

\item{Name}{The name of the preset you are creating.}

\item{Category}{Optional. A category for the preset you are creating.}

\item{Description}{Optional. A description of the preset you are creating.}

\item{Tags}{The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.}
}
\description{
Create a new preset. For information about job templates see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html
}
\examples{

}
