% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{delete_job_template}
\alias{delete_job_template}
\title{Permanently delete a job template you have created}
\usage{
delete_job_template(Name)
}
\arguments{
\item{Name}{[required] The name of the job template to be deleted.}
}
\description{
Permanently delete a job template you have created.
}
\section{Accepted Parameters}{
\preformatted{delete_job_template(
  Name = "string"
)
}
}

