% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{delete_preset}
\alias{delete_preset}
\title{Permanently delete a preset you have created}
\usage{
delete_preset(Name)
}
\arguments{
\item{Name}{The name of the preset to be deleted.}
}
\description{
Permanently delete a preset you have created.
}
\examples{

}
