% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{delete_queue}
\alias{delete_queue}
\title{Permanently delete a queue you have created}
\usage{
delete_queue(Name)
}
\arguments{
\item{Name}{The name of the queue to be deleted.}
}
\description{
Permanently delete a queue you have created.
}
\examples{

}
