% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{disassociate_certificate}
\alias{disassociate_certificate}
\title{Removes an association between the Amazon Resource Name (ARN) of an AWS Certificate Manager (ACM) certificate and an AWS Elemental MediaConvert resource}
\usage{
disassociate_certificate(
  Arn = "string"
)
}
\arguments{
\item{Arn}{[required] The ARN of the ACM certificate that you want to disassociate from your MediaConvert resource.}
}
\description{
Removes an association between the Amazon Resource Name (ARN) of an AWS Certificate Manager (ACM) certificate and an AWS Elemental MediaConvert resource.
}
\examples{

}
