% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{get_job}
\alias{get_job}
\title{Retrieve the JSON for a specific completed transcoding job}
\usage{
get_job(
  Id = "string"
)
}
\arguments{
\item{Id}{[required] the job ID of the job.}
}
\description{
Retrieve the JSON for a specific completed transcoding job.
}
\examples{

}
