% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{get_job_template}
\alias{get_job_template}
\title{Retrieve the JSON for a specific job template}
\usage{
get_job_template(Name)
}
\arguments{
\item{Name}{[required] The name of the job template.}
}
\description{
Retrieve the JSON for a specific job template.
}
\section{Accepted Parameters}{
\preformatted{get_job_template(
  Name = "string"
)
}
}

