% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{get_preset}
\alias{get_preset}
\title{Retrieve the JSON for a specific preset}
\usage{
get_preset(Name)
}
\arguments{
\item{Name}{[required] The name of the preset.}
}
\description{
Retrieve the JSON for a specific preset.
}
\section{Accepted Parameters}{
\preformatted{get_preset(
  Name = "string"
)
}
}

