% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{get_queue}
\alias{get_queue}
\title{Retrieve the JSON for a specific queue}
\usage{
get_queue(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The name of the queue that you want information about.}
}
\description{
Retrieve the JSON for a specific queue.
}
\examples{

}
