% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{list_job_templates}
\alias{list_job_templates}
\title{Retrieve a JSON array of up to twenty of your job templates}
\usage{
list_job_templates(Category = NULL, ListBy = NULL, MaxResults = NULL,
  NextToken = NULL, Order = NULL)
}
\arguments{
\item{Category}{Optionally, specify a job template category to limit responses to only job templates from that category.}

\item{ListBy}{}

\item{MaxResults}{Optional. Number of job templates, up to twenty, that will be returned at one time.}

\item{NextToken}{Use this string, provided with the response to a previous request, to request the next batch of job templates.}

\item{Order}{}
}
\description{
Retrieve a JSON array of up to twenty of your job templates. This will return the templates themselves, not just a list of them. To retrieve the next twenty templates, use the nextToken string returned with the array
}
\examples{

}
