% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{list_presets}
\alias{list_presets}
\title{Retrieve a JSON array of up to twenty of your presets}
\usage{
list_presets(Category = NULL, ListBy = NULL, MaxResults = NULL,
  NextToken = NULL, Order = NULL)
}
\arguments{
\item{Category}{Optionally, specify a preset category to limit responses to only presets from that category.}

\item{ListBy}{}

\item{MaxResults}{Optional. Number of presets, up to twenty, that will be returned at one time}

\item{NextToken}{Use this string, provided with the response to a previous request, to request the next batch of presets.}

\item{Order}{}
}
\description{
Retrieve a JSON array of up to twenty of your presets. This will return the presets themselves, not just a list of them. To retrieve the next twenty presets, use the nextToken string returned with the array.
}
\section{Accepted Parameters}{
\preformatted{list_presets(
  Category = "string",
  ListBy = "NAME"|"CREATION_DATE"|"SYSTEM",
  MaxResults = 123,
  NextToken = "string",
  Order = "ASCENDING"|"DESCENDING"
)
}
}

