% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_package.R
\docType{package}
\name{paws.mediaconvert-package}
\alias{paws.mediaconvert}
\alias{paws.mediaconvert-package}
\title{paws.mediaconvert: AWS Elemental MediaConvert}
\description{
AWS Elemental MediaConvert
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
