% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{tag_resource}
\alias{tag_resource}
\title{Add tags to a MediaConvert queue, preset, or job template}
\usage{
tag_resource(Arn, Tags)
}
\arguments{
\item{Arn}{The Amazon Resource Name (ARN) of the resource that you want to tag. To get the ARN, send a GET request with the resource name.}

\item{Tags}{The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.}
}
\description{
Add tags to a MediaConvert queue, preset, or job template. For information about tagging, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/tagging-resources.html
}
\examples{

}
