% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{update_job_template}
\alias{update_job_template}
\title{Modify one of your existing job templates}
\usage{
update_job_template(Name, Category = NULL, Description = NULL,
  Queue = NULL, Settings = NULL)
}
\arguments{
\item{Name}{The name of the job template you are modifying}

\item{Category}{The new category for the job template, if you are changing it.}

\item{Description}{The new description for the job template, if you are changing it.}

\item{Queue}{The new queue for the job template, if you are changing it.}

\item{Settings}{}
}
\description{
Modify one of your existing job templates.
}
\examples{

}
