% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{update_preset}
\alias{update_preset}
\title{Modify one of your existing presets}
\usage{
update_preset(Name, Category = NULL, Description = NULL, Settings = NULL)
}
\arguments{
\item{Name}{The name of the preset you are modifying.}

\item{Category}{The new category for the preset, if you are changing it.}

\item{Description}{The new description for the preset, if you are changing it.}

\item{Settings}{}
}
\description{
Modify one of your existing presets.
}
\examples{

}
