% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{update_queue}
\alias{update_queue}
\title{Modify one of your existing queues}
\usage{
update_queue(Description = NULL, Name, ReservationPlanSettings = NULL,
  Status = NULL)
}
\arguments{
\item{Description}{The new description for the queue, if you are changing it.}

\item{Name}{[required] The name of the queue that you are modifying.}

\item{ReservationPlanSettings}{Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.}

\item{Status}{Pause or activate a queue by changing its status between ACTIVE and PAUSED. If you pause a queue, jobs in that queue won't begin. Jobs that are running when you pause the queue continue to run until they finish or result in an error.}
}
\description{
Modify one of your existing queues.
}
\section{Accepted Parameters}{
\preformatted{update_queue(
  Description = "string",
  Name = "string",
  ReservationPlanSettings = list(
    Commitment = "ONE_YEAR",
    RenewalType = "AUTO_RENEW"|"EXPIRE",
    ReservedSlots = 123
  ),
  Status = "ACTIVE"|"PAUSED"
)
}
}

