% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{create_queue}
\alias{create_queue}
\title{Create a new transcoding queue}
\usage{
create_queue(Description = NULL, Name, PricingPlan = NULL,
  ReservationPlanSettings = NULL, Tags = NULL)
}
\arguments{
\item{Description}{Optional. A description of the queue that you are creating.}

\item{Name}{[required] The name of the queue that you are creating.}

\item{PricingPlan}{Specifies whether the pricing plan for the queue is on-demand or reserved. For on-demand, you pay per minute, billed in increments of .01 minute. For reserved, you pay for the transcoding capacity of the entire queue, regardless of how much or how little you use it. Reserved pricing requires a 12-month commitment. When you use the API to create a queue, the default is on-demand.}

\item{ReservationPlanSettings}{Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.}

\item{Tags}{The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.}
}
\description{
Create a new transcoding queue. For information about queues, see Working With Queues in the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html
}
\section{Accepted Parameters}{
\preformatted{create_queue(
  Description = "string",
  Name = "string",
  PricingPlan = "ON_DEMAND"|"RESERVED",
  ReservationPlanSettings = list(
    Commitment = "ONE_YEAR",
    RenewalType = "AUTO_RENEW"|"EXPIRE",
    ReservedSlots = 123
  ),
  Tags = list(
    "string"
  )
)
}
}

