% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{update_preset}
\alias{update_preset}
\title{Modify one of your existing presets}
\usage{
update_preset(Category = NULL, Description = NULL, Name,
  Settings = NULL)
}
\arguments{
\item{Category}{The new category for the preset, if you are changing it.}

\item{Description}{The new description for the preset, if you are changing it.}

\item{Name}{[required] The name of the preset you are modifying.}

\item{Settings}{}
}
\description{
Modify one of your existing presets.
}
\section{Accepted Parameters}{
\preformatted{update_preset(
  Category = "string",
  Description = "string",
  Name = "string",
  Settings = list(
    AudioDescriptions = list(
      list(
        AudioNormalizationSettings = list(
          Algorithm = "ITU_BS_1770_1"|"ITU_BS_1770_2",
          AlgorithmControl = "CORRECT_AUDIO"|"MEASURE_ONLY",
          CorrectionGateLevel = 123,
          LoudnessLogging = "LOG"|"DONT_LOG",
          PeakCalculation = "TRUE_PEAK"|"NONE",
          TargetLkfs = 123.0
        ),
        AudioSourceName = "string",
        AudioType = 123,
        AudioTypeControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
        CodecSettings = list(
          AacSettings = list(
            AudioDescriptionBroadcasterMix = "BROADCASTER_MIXED_AD"|"NORMAL",
            Bitrate = 123,
            CodecProfile = "LC"|"HEV1"|"HEV2",
            CodingMode = "AD_RECEIVER_MIX"|"CODING_MODE_1_0"|"CODING_MODE_1_1"|"CODING_MODE_2_0"|"CODING_MODE_5_1",
            RateControlMode = "CBR"|"VBR",
            RawFormat = "LATM_LOAS"|"NONE",
            SampleRate = 123,
            Specification = "MPEG2"|"MPEG4",
            VbrQuality = "LOW"|"MEDIUM_LOW"|"MEDIUM_HIGH"|"HIGH"
          ),
          Ac3Settings = list(
            Bitrate = 123,
            BitstreamMode = "COMPLETE_MAIN"|"COMMENTARY"|"DIALOGUE"|"EMERGENCY"|"HEARING_IMPAIRED"|"MUSIC_AND_EFFECTS"|"VISUALLY_IMPAIRED"|"VOICE_OVER",
            CodingMode = "CODING_MODE_1_0"|"CODING_MODE_1_1"|"CODING_MODE_2_0"|"CODING_MODE_3_2_LFE",
            Dialnorm = 123,
            DynamicRangeCompressionProfile = "FILM_STANDARD"|"NONE",
            LfeFilter = "ENABLED"|"DISABLED",
            MetadataControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
            SampleRate = 123
          ),
          AiffSettings = list(
            BitDepth = 123,
            Channels = 123,
            SampleRate = 123
          ),
          Codec = "AAC"|"MP2"|"WAV"|"AIFF"|"AC3"|"EAC3"|"PASSTHROUGH",
          Eac3Settings = list(
            AttenuationControl = "ATTENUATE_3_DB"|"NONE",
            Bitrate = 123,
            BitstreamMode = "COMPLETE_MAIN"|"COMMENTARY"|"EMERGENCY"|"HEARING_IMPAIRED"|"VISUALLY_IMPAIRED",
            CodingMode = "CODING_MODE_1_0"|"CODING_MODE_2_0"|"CODING_MODE_3_2",
            DcFilter = "ENABLED"|"DISABLED",
            Dialnorm = 123,
            DynamicRangeCompressionLine = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
            DynamicRangeCompressionRf = "NONE"|"FILM_STANDARD"|"FILM_LIGHT"|"MUSIC_STANDARD"|"MUSIC_LIGHT"|"SPEECH",
            LfeControl = "LFE"|"NO_LFE",
            LfeFilter = "ENABLED"|"DISABLED",
            LoRoCenterMixLevel = 123.0,
            LoRoSurroundMixLevel = 123.0,
            LtRtCenterMixLevel = 123.0,
            LtRtSurroundMixLevel = 123.0,
            MetadataControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
            PassthroughControl = "WHEN_POSSIBLE"|"NO_PASSTHROUGH",
            PhaseControl = "SHIFT_90_DEGREES"|"NO_SHIFT",
            SampleRate = 123,
            StereoDownmix = "NOT_INDICATED"|"LO_RO"|"LT_RT"|"DPL2",
            SurroundExMode = "NOT_INDICATED"|"ENABLED"|"DISABLED",
            SurroundMode = "NOT_INDICATED"|"ENABLED"|"DISABLED"
          ),
          Mp2Settings = list(
            Bitrate = 123,
            Channels = 123,
            SampleRate = 123
          ),
          WavSettings = list(
            BitDepth = 123,
            Channels = 123,
            Format = "RIFF"|"RF64",
            SampleRate = 123
          )
        ),
        CustomLanguageCode = "string",
        LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
        LanguageCodeControl = "FOLLOW_INPUT"|"USE_CONFIGURED",
        RemixSettings = list(
          ChannelMapping = list(
            OutputChannels = list(
              list(
                InputChannels = list(
                  123
                )
              )
            )
          ),
          ChannelsIn = 123,
          ChannelsOut = 123
        ),
        StreamName = "string"
      )
    ),
    CaptionDescriptions = list(
      list(
        CustomLanguageCode = "string",
        DestinationSettings = list(
          BurninDestinationSettings = list(
            Alignment = "CENTERED"|"LEFT",
            BackgroundColor = "NONE"|"BLACK"|"WHITE",
            BackgroundOpacity = 123,
            FontColor = "WHITE"|"BLACK"|"YELLOW"|"RED"|"GREEN"|"BLUE",
            FontOpacity = 123,
            FontResolution = 123,
            FontSize = 123,
            OutlineColor = "BLACK"|"WHITE"|"YELLOW"|"RED"|"GREEN"|"BLUE",
            OutlineSize = 123,
            ShadowColor = "NONE"|"BLACK"|"WHITE",
            ShadowOpacity = 123,
            ShadowXOffset = 123,
            ShadowYOffset = 123,
            TeletextSpacing = "FIXED_GRID"|"PROPORTIONAL",
            XPosition = 123,
            YPosition = 123
          ),
          DestinationType = "BURN_IN"|"DVB_SUB"|"EMBEDDED"|"SCC"|"SRT"|"TELETEXT"|"TTML"|"WEBVTT",
          DvbSubDestinationSettings = list(
            Alignment = "CENTERED"|"LEFT",
            BackgroundColor = "NONE"|"BLACK"|"WHITE",
            BackgroundOpacity = 123,
            FontColor = "WHITE"|"BLACK"|"YELLOW"|"RED"|"GREEN"|"BLUE",
            FontOpacity = 123,
            FontResolution = 123,
            FontSize = 123,
            OutlineColor = "BLACK"|"WHITE"|"YELLOW"|"RED"|"GREEN"|"BLUE",
            OutlineSize = 123,
            ShadowColor = "NONE"|"BLACK"|"WHITE",
            ShadowOpacity = 123,
            ShadowXOffset = 123,
            ShadowYOffset = 123,
            TeletextSpacing = "FIXED_GRID"|"PROPORTIONAL",
            XPosition = 123,
            YPosition = 123
          ),
          SccDestinationSettings = list(
            Framerate = "FRAMERATE_23_97"|"FRAMERATE_24"|"FRAMERATE_29_97_DROPFRAME"|"FRAMERATE_29_97_NON_DROPFRAME"
          ),
          TeletextDestinationSettings = list(
            PageNumber = "string"
          ),
          TtmlDestinationSettings = list(
            StylePassthrough = "ENABLED"|"DISABLED"
          )
        ),
        LanguageCode = "ENG"|"SPA"|"FRA"|"DEU"|"GER"|"ZHO"|"ARA"|"HIN"|"JPN"|"RUS"|"POR"|"ITA"|"URD"|"VIE"|"KOR"|"PAN"|"ABK"|"AAR"|"AFR"|"AKA"|"SQI"|"AMH"|"ARG"|"HYE"|"ASM"|"AVA"|"AVE"|"AYM"|"AZE"|"BAM"|"BAK"|"EUS"|"BEL"|"BEN"|"BIH"|"BIS"|"BOS"|"BRE"|"BUL"|"MYA"|"CAT"|"KHM"|"CHA"|"CHE"|"NYA"|"CHU"|"CHV"|"COR"|"COS"|"CRE"|"HRV"|"CES"|"DAN"|"DIV"|"NLD"|"DZO"|"ENM"|"EPO"|"EST"|"EWE"|"FAO"|"FIJ"|"FIN"|"FRM"|"FUL"|"GLA"|"GLG"|"LUG"|"KAT"|"ELL"|"GRN"|"GUJ"|"HAT"|"HAU"|"HEB"|"HER"|"HMO"|"HUN"|"ISL"|"IDO"|"IBO"|"IND"|"INA"|"ILE"|"IKU"|"IPK"|"GLE"|"JAV"|"KAL"|"KAN"|"KAU"|"KAS"|"KAZ"|"KIK"|"KIN"|"KIR"|"KOM"|"KON"|"KUA"|"KUR"|"LAO"|"LAT"|"LAV"|"LIM"|"LIN"|"LIT"|"LUB"|"LTZ"|"MKD"|"MLG"|"MSA"|"MAL"|"MLT"|"GLV"|"MRI"|"MAR"|"MAH"|"MON"|"NAU"|"NAV"|"NDE"|"NBL"|"NDO"|"NEP"|"SME"|"NOR"|"NOB"|"NNO"|"OCI"|"OJI"|"ORI"|"ORM"|"OSS"|"PLI"|"FAS"|"POL"|"PUS"|"QUE"|"QAA"|"RON"|"ROH"|"RUN"|"SMO"|"SAG"|"SAN"|"SRD"|"SRB"|"SNA"|"III"|"SND"|"SIN"|"SLK"|"SLV"|"SOM"|"SOT"|"SUN"|"SWA"|"SSW"|"SWE"|"TGL"|"TAH"|"TGK"|"TAM"|"TAT"|"TEL"|"THA"|"BOD"|"TIR"|"TON"|"TSO"|"TSN"|"TUR"|"TUK"|"TWI"|"UIG"|"UKR"|"UZB"|"VEN"|"VOL"|"WLN"|"CYM"|"FRY"|"WOL"|"XHO"|"YID"|"YOR"|"ZHA"|"ZUL"|"ORJ"|"QPC"|"TNG",
        LanguageDescription = "string"
      )
    ),
    ContainerSettings = list(
      Container = "F4V"|"ISMV"|"M2TS"|"M3U8"|"CMFC"|"MOV"|"MP4"|"MPD"|"MXF"|"RAW",
      F4vSettings = list(
        MoovPlacement = "PROGRESSIVE_DOWNLOAD"|"NORMAL"
      ),
      M2tsSettings = list(
        AudioBufferModel = "DVB"|"ATSC",
        AudioFramesPerPes = 123,
        AudioPids = list(
          123
        ),
        Bitrate = 123,
        BufferModel = "MULTIPLEX"|"NONE",
        DvbNitSettings = list(
          NetworkId = 123,
          NetworkName = "string",
          NitInterval = 123
        ),
        DvbSdtSettings = list(
          OutputSdt = "SDT_FOLLOW"|"SDT_FOLLOW_IF_PRESENT"|"SDT_MANUAL"|"SDT_NONE",
          SdtInterval = 123,
          ServiceName = "string",
          ServiceProviderName = "string"
        ),
        DvbSubPids = list(
          123
        ),
        DvbTdtSettings = list(
          TdtInterval = 123
        ),
        DvbTeletextPid = 123,
        EbpAudioInterval = "VIDEO_AND_FIXED_INTERVALS"|"VIDEO_INTERVAL",
        EbpPlacement = "VIDEO_AND_AUDIO_PIDS"|"VIDEO_PID",
        EsRateInPes = "INCLUDE"|"EXCLUDE",
        FragmentTime = 123.0,
        MaxPcrInterval = 123,
        MinEbpInterval = 123,
        NielsenId3 = "INSERT"|"NONE",
        NullPacketBitrate = 123.0,
        PatInterval = 123,
        PcrControl = "PCR_EVERY_PES_PACKET"|"CONFIGURED_PCR_PERIOD",
        PcrPid = 123,
        PmtInterval = 123,
        PmtPid = 123,
        PrivateMetadataPid = 123,
        ProgramNumber = 123,
        RateMode = "VBR"|"CBR",
        Scte35Pid = 123,
        Scte35Source = "PASSTHROUGH"|"NONE",
        SegmentationMarkers = "NONE"|"RAI_SEGSTART"|"RAI_ADAPT"|"PSI_SEGSTART"|"EBP"|"EBP_LEGACY",
        SegmentationStyle = "MAINTAIN_CADENCE"|"RESET_CADENCE",
        SegmentationTime = 123.0,
        TimedMetadataPid = 123,
        TransportStreamId = 123,
        VideoPid = 123
      ),
      M3u8Settings = list(
        AudioFramesPerPes = 123,
        AudioPids = list(
          123
        ),
        NielsenId3 = "INSERT"|"NONE",
        PatInterval = 123,
        PcrControl = "PCR_EVERY_PES_PACKET"|"CONFIGURED_PCR_PERIOD",
        PcrPid = 123,
        PmtInterval = 123,
        PmtPid = 123,
        PrivateMetadataPid = 123,
        ProgramNumber = 123,
        Scte35Pid = 123,
        Scte35Source = "PASSTHROUGH"|"NONE",
        TimedMetadata = "PASSTHROUGH"|"NONE",
        TimedMetadataPid = 123,
        TransportStreamId = 123,
        VideoPid = 123
      ),
      MovSettings = list(
        ClapAtom = "INCLUDE"|"EXCLUDE",
        CslgAtom = "INCLUDE"|"EXCLUDE",
        Mpeg2FourCCControl = "XDCAM"|"MPEG",
        PaddingControl = "OMNEON"|"NONE",
        Reference = "SELF_CONTAINED"|"EXTERNAL"
      ),
      Mp4Settings = list(
        CslgAtom = "INCLUDE"|"EXCLUDE",
        FreeSpaceBox = "INCLUDE"|"EXCLUDE",
        MoovPlacement = "PROGRESSIVE_DOWNLOAD"|"NORMAL",
        Mp4MajorBrand = "string"
      )
    ),
    VideoDescription = list(
      AfdSignaling = "NONE"|"AUTO"|"FIXED",
      AntiAlias = "DISABLED"|"ENABLED",
      CodecSettings = list(
        Codec = "FRAME_CAPTURE"|"H_264"|"H_265"|"MPEG2"|"PRORES",
        FrameCaptureSettings = list(
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          MaxCaptures = 123,
          Quality = 123
        ),
        H264Settings = list(
          AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
          Bitrate = 123,
          CodecLevel = "AUTO"|"LEVEL_1"|"LEVEL_1_1"|"LEVEL_1_2"|"LEVEL_1_3"|"LEVEL_2"|"LEVEL_2_1"|"LEVEL_2_2"|"LEVEL_3"|"LEVEL_3_1"|"LEVEL_3_2"|"LEVEL_4"|"LEVEL_4_1"|"LEVEL_4_2"|"LEVEL_5"|"LEVEL_5_1"|"LEVEL_5_2",
          CodecProfile = "BASELINE"|"HIGH"|"HIGH_10BIT"|"HIGH_422"|"HIGH_422_10BIT"|"MAIN",
          DynamicSubGop = "ADAPTIVE"|"STATIC",
          EntropyEncoding = "CABAC"|"CAVLC",
          FieldEncoding = "PAFF"|"FORCE_FIELD",
          FlickerAdaptiveQuantization = "DISABLED"|"ENABLED",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopBReference = "DISABLED"|"ENABLED",
          GopClosedCadence = 123,
          GopSize = 123.0,
          GopSizeUnits = "FRAMES"|"SECONDS",
          HrdBufferInitialFillPercentage = 123,
          HrdBufferSize = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          MaxBitrate = 123,
          MinIInterval = 123,
          NumberBFramesBetweenReferenceFrames = 123,
          NumberReferenceFrames = 123,
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          QualityTuningLevel = "SINGLE_PASS"|"SINGLE_PASS_HQ"|"MULTI_PASS_HQ",
          QvbrSettings = list(
            MaxAverageBitrate = 123,
            QvbrQualityLevel = 123
          ),
          RateControlMode = "VBR"|"CBR"|"QVBR",
          RepeatPps = "DISABLED"|"ENABLED",
          SceneChangeDetect = "DISABLED"|"ENABLED",
          Slices = 123,
          SlowPal = "DISABLED"|"ENABLED",
          Softness = 123,
          SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
          Syntax = "DEFAULT"|"RP2027",
          Telecine = "NONE"|"SOFT"|"HARD",
          TemporalAdaptiveQuantization = "DISABLED"|"ENABLED",
          UnregisteredSeiTimecode = "DISABLED"|"ENABLED"
        ),
        H265Settings = list(
          AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH"|"HIGHER"|"MAX",
          AlternateTransferFunctionSei = "DISABLED"|"ENABLED",
          Bitrate = 123,
          CodecLevel = "AUTO"|"LEVEL_1"|"LEVEL_2"|"LEVEL_2_1"|"LEVEL_3"|"LEVEL_3_1"|"LEVEL_4"|"LEVEL_4_1"|"LEVEL_5"|"LEVEL_5_1"|"LEVEL_5_2"|"LEVEL_6"|"LEVEL_6_1"|"LEVEL_6_2",
          CodecProfile = "MAIN_MAIN"|"MAIN_HIGH"|"MAIN10_MAIN"|"MAIN10_HIGH"|"MAIN_422_8BIT_MAIN"|"MAIN_422_8BIT_HIGH"|"MAIN_422_10BIT_MAIN"|"MAIN_422_10BIT_HIGH",
          DynamicSubGop = "ADAPTIVE"|"STATIC",
          FlickerAdaptiveQuantization = "DISABLED"|"ENABLED",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopBReference = "DISABLED"|"ENABLED",
          GopClosedCadence = 123,
          GopSize = 123.0,
          GopSizeUnits = "FRAMES"|"SECONDS",
          HrdBufferInitialFillPercentage = 123,
          HrdBufferSize = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          MaxBitrate = 123,
          MinIInterval = 123,
          NumberBFramesBetweenReferenceFrames = 123,
          NumberReferenceFrames = 123,
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          QualityTuningLevel = "SINGLE_PASS"|"SINGLE_PASS_HQ"|"MULTI_PASS_HQ",
          QvbrSettings = list(
            MaxAverageBitrate = 123,
            QvbrQualityLevel = 123
          ),
          RateControlMode = "VBR"|"CBR"|"QVBR",
          SampleAdaptiveOffsetFilterMode = "DEFAULT"|"ADAPTIVE"|"OFF",
          SceneChangeDetect = "DISABLED"|"ENABLED",
          Slices = 123,
          SlowPal = "DISABLED"|"ENABLED",
          SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
          Telecine = "NONE"|"SOFT"|"HARD",
          TemporalAdaptiveQuantization = "DISABLED"|"ENABLED",
          TemporalIds = "DISABLED"|"ENABLED",
          Tiles = "DISABLED"|"ENABLED",
          UnregisteredSeiTimecode = "DISABLED"|"ENABLED",
          WriteMp4PackagingType = "HVC1"|"HEV1"
        ),
        Mpeg2Settings = list(
          AdaptiveQuantization = "OFF"|"LOW"|"MEDIUM"|"HIGH",
          Bitrate = 123,
          CodecLevel = "AUTO"|"LOW"|"MAIN"|"HIGH1440"|"HIGH",
          CodecProfile = "MAIN"|"PROFILE_422",
          DynamicSubGop = "ADAPTIVE"|"STATIC",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          GopClosedCadence = 123,
          GopSize = 123.0,
          GopSizeUnits = "FRAMES"|"SECONDS",
          HrdBufferInitialFillPercentage = 123,
          HrdBufferSize = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          IntraDcPrecision = "AUTO"|"INTRA_DC_PRECISION_8"|"INTRA_DC_PRECISION_9"|"INTRA_DC_PRECISION_10"|"INTRA_DC_PRECISION_11",
          MaxBitrate = 123,
          MinIInterval = 123,
          NumberBFramesBetweenReferenceFrames = 123,
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          QualityTuningLevel = "SINGLE_PASS"|"MULTI_PASS",
          RateControlMode = "VBR"|"CBR",
          SceneChangeDetect = "DISABLED"|"ENABLED",
          SlowPal = "DISABLED"|"ENABLED",
          Softness = 123,
          SpatialAdaptiveQuantization = "DISABLED"|"ENABLED",
          Syntax = "DEFAULT"|"D_10",
          Telecine = "NONE"|"SOFT"|"HARD",
          TemporalAdaptiveQuantization = "DISABLED"|"ENABLED"
        ),
        ProresSettings = list(
          CodecProfile = "APPLE_PRORES_422"|"APPLE_PRORES_422_HQ"|"APPLE_PRORES_422_LT"|"APPLE_PRORES_422_PROXY",
          FramerateControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          FramerateConversionAlgorithm = "DUPLICATE_DROP"|"INTERPOLATE",
          FramerateDenominator = 123,
          FramerateNumerator = 123,
          InterlaceMode = "PROGRESSIVE"|"TOP_FIELD"|"BOTTOM_FIELD"|"FOLLOW_TOP_FIELD"|"FOLLOW_BOTTOM_FIELD",
          ParControl = "INITIALIZE_FROM_SOURCE"|"SPECIFIED",
          ParDenominator = 123,
          ParNumerator = 123,
          SlowPal = "DISABLED"|"ENABLED",
          Telecine = "NONE"|"HARD"
        )
      ),
      ColorMetadata = "IGNORE"|"INSERT",
      Crop = list(
        Height = 123,
        Width = 123,
        X = 123,
        Y = 123
      ),
      DropFrameTimecode = "DISABLED"|"ENABLED",
      FixedAfd = 123,
      Height = 123,
      Position = list(
        Height = 123,
        Width = 123,
        X = 123,
        Y = 123
      ),
      RespondToAfd = "NONE"|"RESPOND"|"PASSTHROUGH",
      ScalingBehavior = "DEFAULT"|"STRETCH_TO_OUTPUT",
      Sharpness = 123,
      TimecodeInsertion = "DISABLED"|"PIC_TIMING_SEI",
      VideoPreprocessors = list(
        ColorCorrector = list(
          Brightness = 123,
          ColorSpaceConversion = "NONE"|"FORCE_601"|"FORCE_709"|"FORCE_HDR10"|"FORCE_HLG_2020",
          Contrast = 123,
          Hdr10Metadata = list(
            BluePrimaryX = 123,
            BluePrimaryY = 123,
            GreenPrimaryX = 123,
            GreenPrimaryY = 123,
            MaxContentLightLevel = 123,
            MaxFrameAverageLightLevel = 123,
            MaxLuminance = 123,
            MinLuminance = 123,
            RedPrimaryX = 123,
            RedPrimaryY = 123,
            WhitePointX = 123,
            WhitePointY = 123
          ),
          Hue = 123,
          Saturation = 123
        ),
        Deinterlacer = list(
          Algorithm = "INTERPOLATE"|"INTERPOLATE_TICKER"|"BLEND"|"BLEND_TICKER",
          Control = "FORCE_ALL_FRAMES"|"NORMAL",
          Mode = "DEINTERLACE"|"INVERSE_TELECINE"|"ADAPTIVE"
        ),
        ImageInserter = list(
          InsertableImages = list(
            list(
              Duration = 123,
              FadeIn = 123,
              FadeOut = 123,
              Height = 123,
              ImageInserterInput = "string",
              ImageX = 123,
              ImageY = 123,
              Layer = 123,
              Opacity = 123,
              StartTime = "string",
              Width = 123
            )
          )
        ),
        NoiseReducer = list(
          Filter = "BILATERAL"|"MEAN"|"GAUSSIAN"|"LANCZOS"|"SHARPEN"|"CONSERVE"|"SPATIAL",
          FilterSettings = list(
            Strength = 123
          ),
          SpatialFilterSettings = list(
            PostFilterSharpenStrength = 123,
            Speed = 123,
            Strength = 123
          )
        ),
        TimecodeBurnin = list(
          FontSize = 123,
          Position = "TOP_CENTER"|"TOP_LEFT"|"TOP_RIGHT"|"MIDDLE_LEFT"|"MIDDLE_CENTER"|"MIDDLE_RIGHT"|"BOTTOM_LEFT"|"BOTTOM_CENTER"|"BOTTOM_RIGHT",
          Prefix = "string"
        )
      ),
      Width = 123
    )
  )
)
}
}

