% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediaconvert_operations.R
\name{update_queue}
\alias{update_queue}
\title{Modify one of your existing queues}
\usage{
update_queue(Name, Description = NULL, ReservationPlanSettings = NULL,
  Status = NULL)
}
\arguments{
\item{Name}{The name of the queue that you are modifying.}

\item{Description}{The new description for the queue, if you are changing it.}

\item{ReservationPlanSettings}{Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.}

\item{Status}{Pause or activate a queue by changing its status between ACTIVE and PAUSED. If you pause a queue, jobs in that queue won't begin. Jobs that are running when you pause the queue continue to run until they finish or result in an error.}
}
\description{
Modify one of your existing queues.
}
\examples{

}
